Add files via upload
This is the Block Configuration Page. This is someone else's design in our class but I added the link called "Apply to warehouse" so you are given the option to store the work until you have time to work on it later. Example- you get pulled to work on something else first before being given the opportunity to come back to these blocks.