Add files via upload
This is the Warehouse where all blocks that you are currently or hope to work on in the future as well as flowgraphs that you are currently working on or hope to in the future are stored. From the Warehouse you choose your project and then you are linked to block configuration or flowgraph configuration or can keep completed work in here as well.