Add files via upload
This is an example of a "New" Release Flow graph. It may not be truly new but maybe new to you or someone suggested you try it out. From here you can go to Block Configuration, Flow graph Configuration, return to Warehouse Page or click on picture of home and it will take you back to Navigation Page.